ml-model/requirements.txt
bleach==6.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
joblib==1.4.2
kaggle==1.6.17
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2025.1
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
text-unidecode==1.3
threadpoolctl==3.5.0
tqdm==4.67.1
tzdata==2025.1
urllib3==2.2.3
webencodings==0.5.1
